+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.2.3] - 2019-03-10
+
 ### Changed
 
 - `CeleryBuilder::task_route` now infallible. Error could be raised during the `build` phase instead.